requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
langchainhub
openai
pydantic
python-dotenv
numpy
scipy
sympy
matplotlib
plotly
pandas
requests
tiktoken
rich
streamlit
